Add GitHub Actions CI with GCC/Clang matrix builds and ASan support
- Add 4 GitHub Actions workflows: build, test, lint, format
- Add CI scripts in scripts/ci/ for local testing
- Add pre-commit configuration for black, ruff, clang-format, mypy
- Add act integration for local CI testing via `femu ci` commands
- Add ASan preloading for Python tests (detects GCC vs Clang runtime)
- Fix Clang build warnings:
  - exec_load_store.c: Use AccessSize enum instead of uint8_t
  - test_decoder.cpp: Add [[maybe_unused]] to register constants
  - test_trustzone.cpp: Use uint32_t for loop variable
- Disable CppUTest memory leak detection for Clang+ASan builds
- Update flake.nix with act, pre-commit, build packages

+++ src/arch/armv8m/executor/tests/test_trustzone.cpp
@@ -577,8 +577,8 @@ TEST_GROUP(TZ_Exception)
         exec.mem.get_ptr = mock_mem_get_ptr;
 
         /* Set up vector table */
-        for (int i = 0; i < 256; i++) {
-            uint32_t addr = (uint32_t)(i * 4);
+        for (uint32_t i = 0; i < 256; i++) {
+            uint32_t addr = i * 4;
             uint32_t handler = 0x1001 + i * 0x100;  /* Dummy handlers */
             mock_memory[addr] = (uint8_t)(handler & 0xFF);
             mock_memory[addr + 1] = (uint8_t)((handler >> 8) & 0xFF);
